Don't use the image optimiser when generating plain text
