+++ src/blogmore/markdown/__init__.py
@@ -4,4 +4,8 @@ This sub-module contains all custom Markdown extensions used by BlogMore.
 Any new Markdown extensions should be added here as their own module.
 """
 
+from blogmore.markdown.extensions import create_custom_extensions
+
+__all__ = ["create_custom_extensions"]
+
 ### __init__.py ends here
